backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
networkx==3.3
postgis==1.0.4
psycopg2-binary==2.9.9
scipy==1.14.1
daphne==4.1.2
django-leaflet==0.30.1
python-dotenv==1.0.1
channels-redis==4.2.0
django-redis==5.4.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
